chore: Nesting did not work
Moved all sidebar components into a flat directory
