+++ app/components/sidebar/header_component.html.erb
