+++ app/components/sidebar/section_component.html.erb
