+++ app/components/sidebar/section_component.rb
@@ -1,0 +1,30 @@
+# frozen_string_literal: true
+
+module Sidebar
+  # Aside navigation section
+  # @example
+  #  <%= render SideBar::SectionComponent.new(title: 'Section title') do |section| %>
+  #   <%= section.with_item(title: 'Item title', url: '#') %>
+  #   <%= section.with_item(title: 'Item title', url: '#') %>
+  #  <% end %>
+  class SectionComponent < BaseComponent
+    renders_many :items, Sidebar::ItemComponent
+
+    def initialize(title: nil, separator: false, fill: false, **system_arguments)
+      @title = title
+      @separator = separator
+      @fill = fill
+      @system_arguments = system_arguments
+    end
+
+    def system_arguments
+      @system_arguments.tap do |opts|
+        opts[:tag] = 'ul'
+        opts[:classes] = class_names(
+          @system_arguments[:classes],
+          'pt-1 space-y-1'
+        )
+      end
+    end
+  end
+end
