+++ app/components/sidebar/sidebar_component.rb
@@ -11,9 +11,9 @@ module Sidebar
   #   <% end %>
   #  <% end %>
   class SidebarComponent < Component
-    renders_one :header, Sidebar::Header::HeaderComponent
-    renders_many :sections, Sidebar::Section::SectionComponent
-    renders_many :items, Sidebar::Item::ItemComponent
+    renders_one :header, Sidebar::HeaderComponent
+    renders_many :sections, Sidebar::SectionComponent
+    renders_many :items, Sidebar::ItemComponent
 
     def initialize(**system_arguments)
       @system_arguments = system_arguments
